refactor: http 요청, 응답 패키지 분리
