+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -18,6 +18,10 @@ import java.util.Map;
 import java.util.Optional;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.exception.HttpStatusException;
+import org.apache.coyote.http11.httprequest.HttpMethod;
+import org.apache.coyote.http11.httprequest.HttpRequest;
+import org.apache.coyote.http11.httpresponse.HttpResponse;
+import org.apache.coyote.http11.httpresponse.HttpStatusCode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
