+++ tomcat/src/main/java/org/apache/coyote/http11/httprequest/HttpRequest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httprequest;
 
 import java.io.BufferedReader;
 import java.io.IOException;
