+++ tomcat/src/main/java/org/apache/coyote/http11/httpresponse/HttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httpresponse;
 
 import java.nio.charset.StandardCharsets;
 import java.util.LinkedHashMap;
