+++ tomcat/src/main/java/org/apache/coyote/http11/httpresponse/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.httpresponse;
 
 public enum HttpStatusCode {
